+++ client/components/MainComponent/MainComponent.js
@@ -1,18 +1,6 @@
-import ExpenseView from '../ExpenseView/ExpenseView.js'
-import AboutView from '../AboutView/AboutView.js'
-import StatsView from '../StatsView/StatsView.js'
 import template from './template.js'
 import store from '../../store/store.js'
-
-const routes = [
-	{ path: '/about', component: AboutView },
-	{ path: '/dash', component: ExpenseView },
-	{ path: '/stats', component: StatsView }
-]
-
-const router = new VueRouter({
-	routes
-})
+import router from '../../router/router.js'
 
 const MainComponent = new Vue({
 	el: '#app',
