+++ client/router/router.js
@@ -1,0 +1,7 @@
+import routes from './routes.js'
+
+const router = new VueRouter({
+	routes
+})
+
+export default router
